+++ H4_Poker_Engine/Interfaces/IDeckFactory.cs
@@ -1,6 +1,6 @@
 ﻿using H4_Poker_Engine.Models;
 
-namespace H4_Poker_Engine
+namespace H4_Poker_Engine.Interfaces
 {
     public interface IDeckFactory
     {
