+++ H4_Poker_Engine/Interfaces/IHandEvaluator.cs
@@ -1,6 +1,6 @@
 ﻿using H4_Poker_Engine.Models;
 
-namespace H4_Poker_Engine.PokerLogic
+namespace H4_Poker_Engine.Interfaces
 {
     public interface IHandEvaluator
     {
