+++ H4_Poker_Engine/PokerLogic/HandEvaluator.cs
@@ -1,4 +1,5 @@
-﻿using H4_Poker_Engine.Models;
+﻿using H4_Poker_Engine.Interfaces;
+using H4_Poker_Engine.Models;
 
 namespace H4_Poker_Engine.PokerLogic
 {
